requirement.txt
dash>=1.21.0
dash-core-components>=1.17.1
dash-html-components>=1.1.4
dash-renderer>=1.9.1
paho-mqtt>=1.5.1
pandas>=1.3.3
plotly>=5.3.1